# openapi_8001.json
{"openapi":"3.1.0","info":{"title":"Syno","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Home","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/ask":{"get":{"summary":"Ask Get","operationId":"ask_get_ask_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}},"post":{"summary":"Ask Post","operationId":"ask_post_ask_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_ask_post_ask_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"get":{"summary":"Login Get","operationId":"login_get_login_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}},"post":{"summary":"Login Post","operationId":"login_post_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_post_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/signup":{"get":{"summary":"Signup Get","operationId":"signup_get_signup_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}},"post":{"summary":"Signup Post","operationId":"signup_post_signup_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_signup_post_signup_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/logout":{"post":{"summary":"Logout","operationId":"logout_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/ai":{"get":{"summary":"Ai Settings Get","operationId":"ai_settings_get_ai_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}},"post":{"summary":"Ai Settings Post","operationId":"ai_settings_post_ai_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_ai_settings_post_ai_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ai/test":{"post":{"summary":"Ai Settings Test","operationId":"ai_settings_test_ai_test_post","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/me/prompt":{"get":{"summary":"Me Prompt Get","operationId":"me_prompt_get_me_prompt_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}},"post":{"summary":"Me Prompt Post","operationId":"me_prompt_post_me_prompt_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_me_prompt_post_me_prompt_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/q/{qid}":{"get":{"summary":"Question Detail","operationId":"question_detail_q__qid__get","parameters":[{"name":"qid","in":"path","required":true,"schema":{"type":"integer","title":"Qid"}}],"responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/q/{qid}/regen":{"post":{"summary":"Question Regen","operationId":"question_regen_q__qid__regen_post","parameters":[{"name":"qid","in":"path","required":true,"schema":{"type":"integer","title":"Qid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/q/{qid}/answer/mine":{"post":{"summary":"Question Answer Mine","operationId":"question_answer_mine_q__qid__answer_mine_post","parameters":[{"name":"qid","in":"path","required":true,"schema":{"type":"integer","title":"Qid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/me/personas":{"get":{"summary":"Me Personas Get","operationId":"me_personas_get_me_personas_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}},"post":{"summary":"Me Personas Post","operationId":"me_personas_post_me_personas_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_me_personas_post_me_personas_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/me/personas/{pid}/delete":{"post":{"summary":"Me Personas Delete","operationId":"me_personas_delete_me_personas__pid__delete_post","parameters":[{"name":"pid","in":"path","required":true,"schema":{"type":"integer","title":"Pid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/me/personas/{pid}/toggle":{"post":{"summary":"Me Personas Toggle","operationId":"me_personas_toggle_me_personas__pid__toggle_post","parameters":[{"name":"pid","in":"path","required":true,"schema":{"type":"integer","title":"Pid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/vote":{"post":{"summary":"Vote","operationId":"vote_vote_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_vote_vote_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/comment":{"post":{"summary":"Comment","operationId":"comment_comment_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_comment_comment_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_ai_settings_post_ai_post":{"properties":{"provider":{"type":"string","title":"Provider","default":"fake"},"model":{"type":"string","title":"Model","default":"gpt-4o-mini"},"compat_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Compat Name"},"base_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Base Url"},"api_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Api Key"},"temperature":{"type":"number","title":"Temperature","default":0.4}},"type":"object","title":"Body_ai_settings_post_ai_post"},"Body_ask_post_ask_post":{"properties":{"title":{"type":"string","title":"Title"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"}},"type":"object","required":["title"],"title":"Body_ask_post_ask_post"},"Body_comment_comment_post":{"properties":{"target_type":{"type":"string","title":"Target Type"},"target_id":{"type":"integer","title":"Target Id"},"content":{"type":"string","title":"Content"},"parent_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Id"}},"type":"object","required":["target_type","target_id","content"],"title":"Body_comment_comment_post"},"Body_login_post_login_post":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"Body_login_post_login_post"},"Body_me_personas_post_me_personas_post":{"properties":{"name":{"type":"string","title":"Name"},"prompt":{"type":"string","title":"Prompt","default":""}},"type":"object","required":["name"],"title":"Body_me_personas_post_me_personas_post"},"Body_me_prompt_post_me_prompt_post":{"properties":{"preset":{"type":"string","title":"Preset","default":""}},"type":"object","title":"Body_me_prompt_post_me_prompt_post"},"Body_signup_post_signup_post":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"Body_signup_post_signup_post"},"Body_vote_vote_post":{"properties":{"target_type":{"type":"string","title":"Target Type"},"target_id":{"type":"integer","title":"Target Id"},"value":{"type":"integer","title":"Value"}},"type":"object","required":["target_type","target_id","value"],"title":"Body_vote_vote_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}